golden_full_project/golden_project/requirements/dev.txt
-r base.txt
django-debug-toolbar==4.3.0
factory-boy==3.3.0
pytest-django==4.8.0
pytest-cov==5.0.0
black==24.3.0
flake8==7.0.0
isort==5.13.2
ipython==8.23.0
